renaming duke.exceptions to duke.exception
