+++ src/main/java/duke/handler/Storage.java
@@ -1,6 +1,6 @@
 package duke.handler;
 
-import duke.exceptions.DukeException;
+import duke.exception.DukeException;
 import duke.task.Deadline;
 import duke.task.Event;
 import duke.task.Task;
